dom/tests/mochitest/general/test_stylesheetPI.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=836809
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 836809</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 836809 **/

  var pi = document.createProcessingInstruction('xml-stylesheet', 'href="/tests/SimpleTest/test.css" type="text/css"');

  function checkSheet(e)
  {
    ok("sheet" in pi, "XMLStyleSheetProcessingInstruction should have a sheet property");
    ok(pi.sheet, "XMLStyleSheetProcessingInstruction should have a sheet property");
  }

  pi.addEventListener("load", checkSheet);
  document.insertBefore(pi, document.documentElement);

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=836809">Mozilla Bug 836809</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
